+++ mitiq/shadows/__init__.py
@@ -1,14 +1,12 @@
-from mitiq.shadows.post_processing import (
+from mitiq.shadows.classical_postprocessing import (
     expectation_estimation_shadow,
     snapshot_state,
     shadow_state_reconstruction,
 )
-from mitiq.shadows.computational_basis_measurement import (
-    shadow_measure_with_executor,
-)
-from mitiq.shadows.rotation_gates import (
+from mitiq.shadows.quantum_processing import (
     generate_random_pauli_strings,
     get_rotated_circuits,
+    get_z_basis_measurement,
 )
 from mitiq.shadows.shadows import execute_with_shadows
 from mitiq.shadows.shadows_utils import (
