+++ mitiq/shadows/shadows_utils.py
@@ -6,8 +6,8 @@ import numpy as np
 
 # given error of the fidelity between the true state and the reconstructed state, return the number of measurements needed
 def min_n_total_measurements(epsilon: float, num_qubits: int) -> int:
-    """
-    Calculate the number of measurements required to satisfy the shadow bound for the Pauli measurement scheme.
+    """Calculate the number of measurements required to satisfy the shadow
+    bound for the Pauli measurement scheme.
 
     Args:
         epsilon (float): The error on the estimator.
@@ -23,8 +23,7 @@ def min_n_total_measurements(epsilon: float, num_qubits: int) -> int:
 def calculate_shadow_bound(
     error: float, observables: list, failure_rate: float
 ) -> Tuple[int, int]:
-    """
-    Calculate the shadow bound for the Pauli measurement scheme.
+    """Calculate the shadow bound for the Pauli measurement scheme.
 
     Args:
         error (float): The error on the estimator.
@@ -55,8 +54,7 @@ def calculate_shadow_bound(
 
 
 def operator_2_norm(R: np.ndarray) -> float:
-    """
-    Calculate the operator 2-norm.
+    """Calculate the operator 2-norm.
 
     Args:
         R (array): The operator whose norm we want to calculate.
@@ -70,8 +68,10 @@ def operator_2_norm(R: np.ndarray) -> float:
 
 
 def fidelity(state_vector: np.ndarray, rho: np.ndarray) -> float:
-    r"""
-    Calculate the fidelity $$F(\rho,\sigma)=\mathrm{Tr}\sqrt{\rho^{1/2}\sigma\rho^{1/2}}$$, when $$\rho=|v\rangle\langle v|$$ is a pure state $$F(\rho,\sigma)=\langle v|\sigma|v\rangle$$.
+    r"""Calculate the fidelity
+    $$F(\rho,\sigma)=\mathrm{Tr}\sqrt{\rho^{1/2}\sigma\rho^{1/2}}$$, when
+    $$\rho=|v\rangle\langle v|$$ is a pure state $$F(\rho,\sigma)=\langle
+    v|\sigma|v\rangle$$.
 
     Args:
         state_vector (array): The vector whose norm we want to calculate.
